+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Version 0.1.2 (in development)
 - Removed `pre1` from the preprocessor.
+- Reimplemented several `series!` types.
+- Fixed `form`/`mold` for `binary!` values with an offset.
+- Added `clear` action.
 
 
 # Version 0.1.1